Allow host to be evaluated at request time
Supports setting the `host` configuration option to a block which is
evaluated in the view context. This allows the WebPack host to be
dynamically determined at request-time, which is particularly useful if
you're attempting to view the page over your local network (to test on
mobile, for instance).

```ruby
config.webpack.dev_server.host = proc { request.host }
```